Check in sample EEG data a tool to extract signal data
This patch checks in sample data from the place Koosha suggested:

	https://www.physionet.org/pn4/eegmmidb/

The data from the above place are contained in the "EDF+" format. Each
.edf file contains 64 signals, recorded at 160 samples per second, while
a subject performed a task. A PDF is included which depicts where each
sensor is located.

Since the sensor that BraiNet is supposed to use only supplies a single
signal, I've also created a little script that can extract one of the
signals from an EDF file (using a Python library to do it).